backend/requirements/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
fastapi==0.115.6
ghp-import==2.1.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-glightbox==0.4.0
mkdocs-material==9.5.50
mkdocs-material-extensions==1.3.1
packaging==24.2
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
python-jose
